refactor(internal-cloud-client): add records for internal cloud client requests
